moved join button to navbar
