+++ client/src/components/App.jsx
@@ -79,7 +79,11 @@ export default class App extends React.Component {
     } = this.state;
     return (
       <div className="app-container">
-        <NavBar currentWorkSpaceName={currentWorkSpaceName} />
+        <NavBar
+          currentWorkSpaceName={currentWorkSpaceName}
+          currentWorkSpaceId={currentWorkSpaceId}
+          currentUser={this.props.location.state.username}
+        />
         <Body
           messages={messages}
           workSpaces={workSpaces}
