+++ client/src/components/NavBar.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+const axios = require('axios');
+
 
 import {
   Collapse,
@@ -23,6 +25,7 @@ export default class NavBar extends React.Component {
     this.toggle = this.toggle.bind(this);
     this.state = {
       isOpen: false,
+      joined: false,
     };
   }
   toggle() {
@@ -30,6 +33,21 @@ export default class NavBar extends React.Component {
       isOpen: !this.state.isOpen,
     });
   }
+
+  handleJoinClick(event) {
+    let { currentWorkSpaceId, currentUser } = this.props;
+    this.setState({ joined: !this.state.joined });
+    const body = {
+      ws: currentWorkSpaceId,
+      user: currentUser,
+      action: this.state.joined ? 'drop' : 'add',
+    };
+    console.log(body);
+    axios.post('/workspaces/membership', body)
+      .then(console.log)
+      .catch(console.log);
+  }
+
   render() {
     return (
       <Navbar color="faded" light expand="md">
@@ -39,6 +57,9 @@ export default class NavBar extends React.Component {
         <h3 className="text-center">
           #{this.props.currentWorkSpaceName || 'select a workspace!'}{' '}
         </h3>
+        <button onClick={event => this.handleJoinClick(event)}>
+          {this.state.joined ? 'Leave' : 'Join'}
+        </button>
         <NavbarToggler onClick={this.toggle} />
         <Collapse isOpen={this.state.isOpen} navbar>
           <Nav className="ml-auto" navbar>
